unneeded casts

+++ src/lib/tta_dec.c
@@ -293,7 +293,7 @@ tta_decode_mch(
 	size_t i;
 	uint j;
 
-	for ( i = 0; i < ni32_target; i += (size_t) nchan ){
+	for ( i = 0; i < ni32_target; i += nchan ){
 		if ( r > read_soft_limit ){ break; }
 #ifdef LIBTTAr_OPT_DISABLE_UNROLLED_1CH
 		prev = 0;	// for mono
